+++ ahrs_platform/ahrs_java_pcsim/src/bottomnav/hitherejoe/com/bottomnavigationsample/SensorFusion.java
@@ -1733,7 +1733,7 @@ public class SensorFusion {
         // correct the hard iron estimate for FMATRIXSCALING and the offsets applied (result in uT)
         for (l = CHX; l <= CHZ; l++)
         {
-            ftrV[l] = (float)(ftrV[l] * DEFAULTB + iOffset[l] * 1.0 * fuTPerCount);
+            ftrV[l] = ftrV[l] * DEFAULTB + iOffset[l] * 1.0 * fuTPerCount;
         }
         // correct the geomagnetic field strength B to correct scaling (result in uT)
         ftrB *= DEFAULTB;
